refactor: score와 result 수정 - chessBoard에서 점수 계산 - result는 진영 별 점수를 갖도록 변경
